Use BuildOptions in foundry_kompile
