+++ src/tests/integration/conftest.py
@@ -10,7 +10,7 @@ from pyk.kore.rpc import kore_server
 from pyk.utils import run_process
 
 from kontrol.foundry import Foundry
-from kontrol.kompile import foundry_kompile
+from kontrol.kompile import BuildOptions, foundry_kompile
 
 from .utils import TEST_DATA_DIR, gen_bin_runtime
 
@@ -68,19 +68,23 @@ def foundry(foundry_root_dir: Path | None, tmp_path_factory: TempPathFactory, wo
             run_process(['forge', 'build'], cwd=foundry_root)
 
             foundry_kompile(
+                BuildOptions(
+                    {
+                        'includes': (),
+                        'requires': [
+                            str(TEST_DATA_DIR / 'lemmas.k'),
+                            str(TEST_DATA_DIR / 'cse-lemmas.k'),
+                            str(TEST_DATA_DIR / 'pausability-lemmas.k'),
+                        ],
+                        'imports': [
+                            'LoopsTest:SUM-TO-N-INVARIANT',
+                            'ArithmeticCallTest:CSE-LEMMAS',
+                            'CSETest:CSE-LEMMAS',
+                            'PortalTest:PAUSABILITY-LEMMAS',
+                        ],
+                    }
+                ),
                 foundry=Foundry(foundry_root),
-                includes=(),
-                requires=[
-                    str(TEST_DATA_DIR / 'lemmas.k'),
-                    str(TEST_DATA_DIR / 'cse-lemmas.k'),
-                    str(TEST_DATA_DIR / 'pausability-lemmas.k'),
-                ],
-                imports=[
-                    'LoopsTest:SUM-TO-N-INVARIANT',
-                    'ArithmeticCallTest:CSE-LEMMAS',
-                    'CSETest:CSE-LEMMAS',
-                    'PortalTest:PAUSABILITY-LEMMAS',
-                ],
             )
 
     session_foundry_root = tmp_path_factory.mktemp('foundry')
