+++ src/tests/profiling/test_foundry_prove.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import sys
 from typing import TYPE_CHECKING
 
-from kontrol.kompile import foundry_kompile
+from kontrol.kompile import BuildOptions, foundry_kompile
 from kontrol.options import ProveOptions, RPCOptions
 from kontrol.prove import foundry_prove
 
@@ -25,7 +25,7 @@ def test_foundy_prove(profile: Profiler, no_use_booster: bool, bug_report: BugRe
     foundry = forge_build(TEST_DATA_DIR, foundry_root)
 
     with profile('kompile.prof', sort_keys=('cumtime', 'tottime'), limit=15):
-        foundry_kompile(foundry=foundry, includes=())
+        foundry_kompile(BuildOptions({'includes': ()}), foundry=foundry)
 
     with profile('prove.prof', sort_keys=('cumtime', 'tottime'), limit=100):
         foundry_prove(
